adding react frontend

+++ graphql/resolvers/booking.js
@@ -12,33 +12,41 @@ module.exports = {
             throw e
         }
     },
-    bookEvent: async (args) => {
+    bookEvent: async (args, req) => {
         try {
-            const checkEventExist = await EventModel.findById(args.eventId);
-            if (checkEventExist) {
-                const { _id } = checkEventExist;
-                const booking = new bookingModel({
-                    user: "5eb4708beeb08c13397eeba0",
-                    event: _id
-                });
-                await booking.save();
-                return await bookingModel.findById(booking._id).populate({
-                    path: "user"
-                }).populate({ path: "event" }).lean();
+            if (req.isAuth) {
+                const checkEventExist = await EventModel.findById(args.eventId);
+                if (checkEventExist) {
+                    const { _id } = checkEventExist;
+                    const booking = new bookingModel({
+                        user: req.userId,
+                        event: _id
+                    });
+                    await booking.save();
+                    return await bookingModel.findById(booking._id).populate({
+                        path: "user"
+                    }).populate({ path: "event" }).lean();
+                } else {
+                    throw new Error("Event does not exist")
+                }
             } else {
-                throw new Error("Event does not exist")
+                throw new Error('Authentication required !!')
             }
         } catch (e) {
             throw e;
         }
     },
-    cancelBooking: async (args) => {
+    cancelBooking: async (args, req) => {
         try {
-            const booking = await bookingModel.findById(args.bookingId).populate({
-                path: "event"
-            }).lean();
-            await bookingModel.deleteOne({ _id: args.bookingId });
-            return booking;
+            if (req.isAuth) {
+                const booking = await bookingModel.findById(args.bookingId).populate({
+                    path: "event"
+                }).lean();
+                await bookingModel.deleteOne({ _id: args.bookingId });
+                return booking;
+            } else {
+                throw new Error('Authentication required !!')
+            }
         } catch (e) {
             throw e
         }
